Moved KdTree functors to their own header files. Split KdTree unit tests. Removed the KdTree Dim template argument. KdTree Splitter argument type replaced. Google Code Guide update for output function arguments. Added SpaceWrapper to simplify interfacing with a Space.
