+++ test/pico_tree/CMakeLists.txt
@@ -15,6 +15,7 @@ include_directories(${CMAKE_CURRENT_LIST_DIR})
 set(TEST_TARGET_SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/box_test.cpp
     ${CMAKE_CURRENT_LIST_DIR}/cover_tree_test.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/kd_tree_builder_test.cpp
     ${CMAKE_CURRENT_LIST_DIR}/kd_tree_test.cpp
     ${CMAKE_CURRENT_LIST_DIR}/metric_test.cpp
     ${CMAKE_CURRENT_LIST_DIR}/std_traits_test.cpp
